requirements.txt
pandas
requests
numpy
streamlit
plotly
ipywidgets
Pillow
anywidget
lxml
html5lib
beautifulsoup4
understatapi
selenium
webdriver_manager